Minor bug fixes and add helpful informaiton to the StartubTab.

+++ MainApplication/MainWindow.xaml.cs
@@ -1235,9 +1235,13 @@ namespace ImagingSIMS.MainApplication
 
             if (Workspace.Registry.ClearPluginData)
             {
-                var fileInfo = new DirectoryInfo(
-                    Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), @"ImagingSIMS\plugins"));
-                fileInfo.GetFiles("*", SearchOption.AllDirectories).ToList().ForEach(file => file.Delete());
+                string folderPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), @"ImagingSIMS\plugins");
+
+                if (Directory.Exists(folderPath))
+                {
+                    var fileInfo = new DirectoryInfo(folderPath);
+                    fileInfo.GetFiles("*", SearchOption.AllDirectories).ToList().ForEach(file => file.Delete());
+                }                
             }
 
             //Closes the application with no further action if the application is in Debug mode
@@ -2078,6 +2082,20 @@ namespace ImagingSIMS.MainApplication
                 await window.SetDataAsync(renderVolumes);
                 window.BeginRendering();
             }
+            catch (DllNotFoundException DNFex)
+            {
+                DialogBox.Show(DNFex.Message,
+                    "Download and install the DirectX runtime package at https://www.microsoft.com/en-us/download/confirmation.aspx?id=8109 to install the necessary files.",
+                    "DLL Not Found", DialogBoxIcon.Stop);
+
+                if (window != null)
+                {
+                    window.Close();
+                    window = null;
+                }
+
+                return;
+            }
             catch (Exception ex)
             {
                 DialogBox db = new DialogBox("There was an error creating the 3D rendering.", ex.Message, "Direct3D", DialogBoxIcon.Stop);
@@ -2145,6 +2163,20 @@ namespace ImagingSIMS.MainApplication
 
                 await window.SetDataAsync(hm.CorrectedHeightData, hm.CorrectedColorData);
                 window.BeginRendering();
+            }
+            catch(DllNotFoundException DNFex)
+            {
+                DialogBox.Show(DNFex.Message, 
+                    "Download and install the DirectX runtime package at https://www.microsoft.com/en-us/download/confirmation.aspx?id=8109 to install the necessary files.", 
+                    "DLL Not Found", DialogBoxIcon.Stop);
+
+                if (window != null)
+                {
+                    window.Close();
+                    window = null;
+                }
+
+                return;
             }
             catch (Exception ex)
             {
@@ -2815,10 +2847,11 @@ namespace ImagingSIMS.MainApplication
         private void LoadFusionHigh(object sender, RoutedEventArgs e)
         {
             OpenFileDialog ofd = new OpenFileDialog();
-            ofd.Filter = "Bitmap Images (.bmp)|*.bmp";
+            ofd.Title = "Open High Res Image";
+            ofd.Filter = "Image Files (.bmp, .jpg, .jpeg, .png)|*.bmp;*.jpg;*.jpeg;*.png|" +
+                            "Bitmap Images (.bmp)|*.bmp|JPEG Images (.jpg, .jpeg)|*.jpg;*.jpeg|PNG Images (.png)|*.png";
             ofd.Multiselect = false;
-            Nullable<bool> result = ofd.ShowDialog();
-            if (result != true) return;
+            if (ofd.ShowDialog() != true) return;
 
             BitmapImage src = new BitmapImage();
             src.BeginInit();
@@ -2860,10 +2893,11 @@ namespace ImagingSIMS.MainApplication
         private void LoadFusionLow(object sender, RoutedEventArgs e)
         {
             OpenFileDialog ofd = new OpenFileDialog();
-            ofd.Filter = "Bitmap Images (.bmp)|*.bmp";
+            ofd.Title = "Open Low Res Image";
+            ofd.Filter = "Image Files (.bmp, .jpg, .jpeg, .png)|*.bmp;*.jpg;*.jpeg;*.png|" +
+                            "Bitmap Images (.bmp)|*.bmp|JPEG Images (.jpg, .jpeg)|*.jpg;*.jpeg|PNG Images (.png)|*.png";
             ofd.Multiselect = false;
-            Nullable<bool> result = ofd.ShowDialog();
-            if (result != true) return;
+            if (ofd.ShowDialog() != true) return;
 
             BitmapImage src = new BitmapImage();
             src.BeginInit();
